# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTCAS API",
    "version": "1.0.0",
    "description": "RTCAS Backend API Definition"
  },
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://api.rtcas.in.th" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/ethics-assessments/import": {
      "post": {
        "summary": "Import ethics assessments from Excel",
        "tags": ["Ethics Assessments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Import success" },
          "400": { "description": "Invalid file or data" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/ethics-assessments": {
      "get": {
        "summary": "List ethics assessments",
        "tags": ["Ethics Assessments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete all ethics assessments",
        "tags": ["Ethics Assessments"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Deleted all" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/ethics-assessments/{id}": {
      "delete": {
        "summary": "Delete ethics assessment by id",
        "tags": ["Ethics Assessments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/evaluations/import": {
      "post": {
        "summary": "Import evaluation scores from Excel",
        "tags": ["Evaluations"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Imported" },
          "400": { "description": "Invalid file or data" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/evaluations": {
      "get": {
        "summary": "List evaluation sheets",
        "tags": ["Evaluations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/evaluations/{id}": {
      "get": {
        "summary": "Get evaluation sheet by id",
        "tags": ["Evaluations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Sheet returned" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update evaluation sheet",
        "tags": ["Evaluations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete evaluation sheet",
        "tags": ["Evaluations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/evaluations/template/download": {
      "get": {
        "summary": "Download evaluation Excel template",
        "tags": ["Evaluations"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Template file" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/exam-results/import": {
      "post": {
        "summary": "Import exam results from Excel",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Imported" },
          "400": { "description": "Invalid file or data" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/exam-results": {
      "get": {
        "summary": "List exam results",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete all exam results",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/exam-results/summary": {
      "get": {
        "summary": "Summarize exam results",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Summary returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/exam-results/{id}": {
      "delete": {
        "summary": "Delete exam result by id",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/exam-results/overview": {
      "get": {
        "summary": "Exam results overview",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Overview returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/exam-results/export": {
      "get": {
        "summary": "Export exam results as Excel",
        "tags": ["Exam Results"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Excel file" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  },
...
}